requirements.txt
Django==1.11

dj-database-url==0.4.2
django-smartfields==1.0.9
django-widget-tweaks==1.4.1
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.2.0
jdcal==1.2
mechanize==0.2.5
Pillow==5.0.0
psycopg2-binary==2.7.4
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
requests==2.18.4
Unidecode==1.0.22